LPS-68072 WYSIWYG rename from lowercase
